Add basic ZMQ subscriber.
Make the Game subscribe to game-block-* notifications as sent from the
Xaya daemon.  While the main loop is running, a new thread listens for
notifications, until the main loop is stopped.

Currently, the notifications are just logged -- in the future, they will
be sent to the core Game instance through some listener interface.