Make MainLoop's interface public.
Remove the "friend Game" declaration for MainLoop and instead make its
intended interface public (so that it can be used from Game without
being friends).

This makes it more explicit what the interface is, clarifying the code.

Of course, MainLoop is still not part of the actual public interface for
libxayagame; but this is expressed by it being part of the "internal"
namespace.