Reduce sleeps and timeouts.
This reduces the sleep times in tests (while waiting for ZMQ to start
listening, for instance, and threads to finish their work) as well as
the timeout that is used in the ZMQ subscriber loop.

The new timeout (100ms) should be more than big enough to not yield a
"busy wait", and the test sleeps of 10ms seem also ample for the tests
to work properly.

This reduces the runtime of the full tests from 8s to 1.2s.