Detect chain of Xaya daemon when connecting.
On the initial RPC connection, detect the chain "type" (mainntest,
testnet, regtest) of the Xaya daemon.  Verify that the chain is the
same if we reconnect; otherwise we would really mess up the state if
the connected daemon suddenly changes from mainnet to testnet or
something like that.

The detected chain will later be used by games to set up the initial
state for each network correctly, and also to choose the right storage
location for the database.