Attach Storage and GameLogic to Game instance.
Define a GameLogic interface (for now incomplete) that will be
implemented by actual games to define their rules.

Keep a reference to a Storage and GameLogic implementation in the Game
instance for future use in the actual Game operations.

Attach Storage and a skeleton GameLogic implementation for the RPS game;
for now, the logic just sets an empty game state at predefined genesis
heights for mainnet, testnet and regtest.