Use fake implementation for getblockchaininfo.
Instead of mocking (with GMock) getblockchaininfo in the RPC server used
for unit tests, this adds a fake implementation that returns preset data
(but without verifying the exact order and number of calls made).  This
is more appropriate -- we do not care about the actual calls, that is
an implementation detail.  We just have to provide data that matches a
particular test situation.