Make chain member variable of GameLogic.
The chain can potentially influence not just the initial game state but
also the state transitions (e.g. different fork heights / rules for
testnet vs mainnet).  Thus it makes sense to have the chain as a general
property of the GameLogic instance, rather than pass it in for each
method as an argument.