Return state directly from GetInitialState.
Change GameLogic::GetInitialState from a void method that returns the
game state through an output argument to a function that returns the
state directly (but still uses output arguments for the associated
genesis height and hash).

This makes the method signature at least a bit simpler to use, and will
be consistent with the forward/backwards processing methods.  All of them
return a game state as the "primary" result, so that can well be returned
directly.