Implement game logic for Mover.
Add the ProcessForward and ProcessBackwards methods to the GameLogic
interface, but they are still unused by the main xayagame logic.

However, the Mover example game already implements them, which finishes
the game logic for that game and makes it ready to use in testing of the
main implementation for game-state updates in xayagame.