Add getcurrentstate method for game RPC server.
The Game class has a new GetCurrentJsonState method, which returns the
current game state as well as some meta information as JSON object.
This method is exposed by the game RPC server as "getcurrentstate", and
is meant to be the main external interface for frontends that need to
get the game state.

This is the main implementation for resolving
https://github.com/xaya/libxayagame/issues/1.