Fix catching-up for only detaches.
When the game was catching up with a target block that was in the
ancestry of the current best block (only detaching blocks), then the
state would not be reinitialised correctly after reaching the target.
In this case, it would just "stop" at being catching-up.

This is a somewhat strange case, but it happens genuinely if a block
is invalidated (for instance in integration testing) and the resulting
detach message is the first detach at all (leading to an interpretation
as missed notification).