Implement pruning support in (Memory)Storage.
Store the block height next to the actual undo data in Storage, which
allows to implement also "pruning up to height N" in a new method.

This contains the changes necessary to the Storage interface (and
implementations) for effective pruning support in the Game class itself
(https://github.com/xaya/libxayagame/issues/8).