Implement (and test) pruning in Mover.
Add a new --enable_pruning=N argument to moverd, which can be used to
enable pruning mode.  A new integration test pruning.py tests that the
game runs correctly with N=0 and N=1 (and a reorg in the latter case).

Also includes minor tweaks to xayagametest that were necessary to
implement the pruning test.