Move GameLogic class to own source files.
Instead of defining the GameLogic class/interface in game.hpp/.cpp,
define it in its own source file gamelogic.hpp/.cpp.  This makes the
code more modular and allows for instance the Mover logic to depend
only on gamelogic and not the larger game.

The new file will also be a good place to add the CachingGame subclass
for games without undo logic in the future.