Provide CachingGame for games without undo.
The new class CachingGame is a subclass of GameLogic, which can be used
to implement simple games (or rather, games with a small expected state).
Those games only need to define a forward-processing function, and won't
need any undo logic.

This works by using the old game state itself as the undo data.  As long
as the size is small and pruning is used on the node, that should work
fine at least for simple and easy games.  It can also work well for games
where the game state itself is held by e.g. an external database, and the
state in libxayagame is just a handle / snapshot ID / revision.

Implements https://github.com/xaya/libxayagame/issues/3.