Introduce transaction mechanism for game state.
This adds new functions to the GameLogic interface, which can be used to
(optionally) implement transaction-semantics for changes to the game
state.  In particular, if the GameLogic class holds internal state that
is modified with ProcessForward / ProcessBackwards, then the new functions
allow it to commit or abort changes to the state in a way synchronised
to updates to the core storage instance.

If SetCurrentGameState on the storage fails for some reason (i.e. throws
an exception) and the storage is implemented in such a way that it has not
been modified if that happens, then the internal state of GameLogic can
be made consistent with that.

If the process crashes (or e.g. power fails) between SetCurrentGameState
and committing the transaction, however, things are still bad (but this
is unlikely).  With the upcoming SQLite integration, however, at least
games that use SQLite for internal state *and* to implement
StorageInterface will be able to do all of that in a single database
transaction and be fully atomic / consistent.