Add DefaultMain based on callbacks.
This includes a second DefaultMain variant, which is based on callbacks
for the game-specific logic instead of an instance of GameLogic.  This
may make it easier to implement for simple games, and may also make
interoperability with C# (writing logic in C#) easier to handle.

(Because not a class needs to be wrapped, but just functions.)