Make storage_tests type-parametrised.
The tests in storage_tests.cpp (currently applied to MemoryStorage)
are just testing the basic interface and nothing specific to
MemoryStorage.  In this change, we turn them into a set of
type-parametrised tests, so that they can be applied to other storage
implementations in the future (e.g. the upcoming SQLiteStorage).