Implement SQLite-based storage.
This adds an implementation of StorageInterface based on an SQLite
database (and for that, it adds SQLite as a new dependency).

Note that StorageInterface could be implemented with a no-SQL database
instead, as only key lookups are required.  Hence, while this implements
a persistent storage for https://github.com/xaya/libxayagame/issues/2,
it is not meant to be used as general persistent storage (another one
based on perhaps LMDB will be added for that use).  Instead, it should
be used when games use SQLite anyway for their state, so that a single
database holds everything and can be updated atomically (see
https://github.com/xaya/libxayagame/issues/11).