Enable and test SQLite storage in mover.
Update the DefaultMain so that the new SQLite storage can be configured
to be used instead of MemoryStorage (including basic logic to set up a
data directory for libxayagame-enabled games).

Mover is extended to have new flags for this feature, as well as an
integration test that uses persistent storage and verifies that it works.