Special support for SQLite-based games.
This adds SQLiteGame as new GameLogic subclass, which can be used for
games that keep their state in an SQLite database.  The class manages
the database, making sure that the state kept (and modified) by the
game is consistent with libxayagame's general "game state" mechanism.

SQLiteGame also handles everything related to rollbacks / undo data
automatically; for this, the SQLite session extension is used, and
generated "changesets" are stored as the undo data.

With this, https://github.com/xaya/libxayagame/issues/11 gets resolved.
This is also (one specific) way to implement the more generic
https://github.com/xaya/libxayagame/issues/5.