Move GameTestFixture to testutils.
This moves part of the logic from GameTests (in game_tests.cpp) to
testutils so it can be shared with other tests in the future.

In particular, the functionality that exposes private internals of Game
by being a friend class is moved, as well as the utility methods for
attaching/detaching blocks.