Add initialisation step for StorageInterface.
The new method StorageInterface::Initialise can be overridden if some
storage implementation needs initialisation besides the constructor
(e.g. to open some external stuff).  The function is called when the
storage is attached to a game (or explicitly as needed in tests).

SQLiteStorage makes use of this to open the database only during
initialisation and not construction.  Otherwise, calls to SetupSchema
will never call overriding methods from subclasses, as those are not
constructed yet at the previous call-site from the constructor.