Game test fixture with "simulated blockchain".
This adds GameTestWithBlockchain to the testutils, which is an extended
GameTestFixture that keeps track of a stack of blocks and associated
moves, so that it is easier to attach and detach blocks in a consistent
way.  This allows to simplify some of the game_tests, and will be
especially useful in other tests.

(Not all uses of CallBlockAttach and CallBlockDetach could be replaced
in game_tests, as some of them deliberately test the reaction of Game
to sequence mismatches or request tokens.  But apart from game_tests,
those situations should be very rare, so that most other tests that
involve a Game instance can rely exclusively on the new, simpler
functions for attaching/detaching blocks.)