Explicitly test game syncing in multiple steps.
Add a new test for syncing of Game where multiple game_sendupdate
requests are necessary to bring the state up-to-date.  This verifies
that operation is correct also after -maxgameblockattaches was
introduced in Xaya Core (https://github.com/xaya/xaya/pull/66).

The logic itself is already working in that case, as we never even
explicitly specify a "toblock" in game_sendupdate requests created
from libxayagame.  But we should have a test that verifies it works,
so that there cannot be any regressions in the future.