Change chain from string to enum.
GameLogic exposes the current "chain" the game is running on to the
actual game implementation, so that it may, for instance, implement
different fork heights for testnet vs mainnet (or have other rule
differences).

This value was previously just taken as string from getblockchaininfo
and passed on.  But this means that any changes to Xaya Core will
directly influence and break games, as well as requiring that any
game actually hardcodes the "magic values" of the string.

Here, we change the type to an enum defined by gamelogic.hpp.  This
makes the possible values clearer, so that game implementations can
just depend on the defined enum values.  The translation from a string
happens directly after calling getblockchaininfo, and translation back
to a string for logging and storage-directory names can be done by
a new utility function ChainToString.

This fixes https://github.com/xaya/libxayagame/issues/17.