Move transactions from GameLogic to storage.
The handling for "atomic transactions" should be part of the storage, not
the GameLogic.  As part of the GameLogic, it can specifically be used
to synchronise changes to the core storage and external states that the
game keeps by itself.  But ultimately, transactions are a concept that is
part of the storage and not the game rules.  Thus moving the functions to
StorageInterface makes more sense.

Games can still synchronise external state, of course, but the proper way
to do it is to subclass storage and make it hold both the libxayagame
data and the game-state data.  Then the storage's transaction functions
can be used to make everything consistent.

By having transactions in the storage, they can benefit standard
implementations like the SQLiteStorage and an upcoming LMDB-based one;
not just special games like SQLiteGame.  Even though atomicity is not
strictly necessary for simple games (as the order of adding/releasing
undo data and updating the game state is carefully chosen), real atomic
transactions are still useful and can furthermore be used to avoid
excessive synchronisations to disk while the game is catching up (this
will be done in a follow up).