Batch atomic transactions from Game.
A new utility class TransactionManager is introduced, which manages
started and finished "atomic transactions" for the storage.  It sits
between the Game instance and the storage that actually executes those
transactions, and enables batching of transactions while the game is
still catching up.  This speeds up syncing a lot, since it avoids too
many synchronisations to disk.

This fixes https://github.com/xaya/libxayagame/issues/16.