Add LMDB-based storage implementation.
This creates the new LMDBStorage class, which implements StorageInterface
based on an LMDB database.  This is a no-SQL database, which is likely
more suited and efficient for basic libxayagame storage than, say,
SQLiteStorage for situations where the game itself does not need
SQLite anyway.

With this, https://github.com/xaya/libxayagame/issues/2 can be resolved.