Correctly fail test when exception is caught.
Make sure that xayagametest scripts exit with a non-zero exit code
(marking the test as failed) if an exception is caught, even if it
was not an assertion failure.  Without this, such errors would show
as success in make check.

Also changes the script to use generatetoaddress in Xaya Core rather
than generate, which has been removed recently.