Integrate LMDB storage option with DefaultMain.
Extend DefaultMain, so that "lmdb" is also accepted as storage type
and an LMDBStorage instance is created in that case.  This makes also
mover capable of using LMDB storage.  The persistence test is extended
to run both with "sqlite" and "lmdb" as storage types to verify that
both of them work in a "real-world setting".