Resize LMDB using RetryWithNewTransaction.
Make use of the RetryWithNewTransaction mechanism to resize the LMDB
map if a write operation fails with MDB_MAP_FULL.  In that case, we
throw RetryWithNewTransaction and set a flag that tells us to actually
perform the resize on the next RollbackTransaction call.  This call
is expected to happen as the thrown exception unwinds the stack and
an object like ActiveTransaction is destructed.