Add TryAbortTransaction for TransactionManager.
The new method TryAbortTransaction in the TransactionManager makes
sure that all active transactions on the underlying storage have been
aborted (if there were any).  This is used before calling Clear() on
the storage from game to make sure that no transactions are active
when Clear() is called (as the interface specification says).