Enforce transaction behaviour in MemoryStorage.
Assert that the behaviour with respect to started transactions is
correct when MemoryStorage methods are called.  While the MemoryStorage
itself does not handle transactions in any way, this ensures that the
various tests that use it catch misbehaving code on that front.

This also has a minor performance hit for non-test code that uses the
MemoryStorage, but that should not matter in practice.  First, the
additional operations are very cheap.  Second, real practical
applications are most likely not using MemoryStorage anyway for non-test
and non-development purposes.