Add RetryWithNewTransaction for StorageInterface.
This is a special exception, which may be thrown by storage methods if
they encounter an issue but the Game instance may retry after aborting
the current transaction and restarting a new one.

One potential candidate for using this is LMDB, where the map size needs
to be increased but that can only be done when no transaction is
currently active.

The Game instance is not yet actually handling these exceptions, but
that will be added in a follow-up commit.