Allow DefaultMain to start a TCP RPC server.
This refactors the construction of the JSON-RPC server connector from
DefaultMain a bit.  A new config option specifies what type of server to
start (if any), and it now allows to start a plain TCP socket based one
rather than the HTTP server.  This helps work around issues on Windows.