Move SleepSome to testutils.
The SleepSome function (previously in mainloop_tests.cpp) just sleeps
the current thread for a small amount of time (10ms at the moment), which
is a hack but useful in some test situations to ensure that other threads
have had time to run and start some action.

Moving this to testutils makes the function available to other tests in
the future, in particular game_tests.cpp.