Split general code out of ChatGame.
Refactor the ChatGame (and related helper code) in sqlitegame_tests.cpp,
so that code that is not actually related to the concrete game rules can
be reused.

Also the general test fixture class is now a template, so that it can be
instantiated with different games.

In the future, we want to have another test game for verifying generated
IDs when inserting stuff.  For this, we can then reuse that general code.