Add test for SQLite game with automatic IDs.
Add another SQLite-based test game, which is about inserting stuff
into tables with automatically generated IDs.  It does a basic
verification of those IDs for reorgs and error rollbacks.

For now, though, the IDs are still based on rowids assigned by SQLite.
That the test works is mostly by accident.  In the next step, we will
implement our own way of managing those IDs and change the test game
to use that instead.