Rename GameLogic callbacks to "Internal".
The GameLogic functions that need to be implemented by actual games now
have an "Internal" suffix.  The non-Internal versions are defined by
GameLogic itself and call the Internal ones.

For now, this does not change anything else.  But it will make it possible
for GameLogic to do extra work before calling the internal functions,
namely setting up a Context instance (and perhaps other stuff in the
future as needed).