Add short sleeps to fix waitforchange.
Add a short sleep when stopping the game daemon, so that the RPC server has
time to reply to waitforchange calls after being notified on them.  Without
this, it may be stopped by the game already before it can reply, leading
to errors at the client.

Also introduces a short sleep in the waitforchange.py test itself, to
ensure that the waiter thread has had time to start up.