Allow customisation of RPC server in DefaultMain.
Implement a framework (using a factory interface) that allows callers of
DefaultMain to optionally customise other classes (besides the mandatory
GameLogic instance).  For now, this allows using a custom RPC server
instead of the default GameRpcServer.

With this in place, we can add another DefaultMain function for SQLite
games and then actually use if for SMC and "Project X".