Do not open database in SQLiteGame constructor.
Instead of passing the filename and opening the underlying database
already in the constructor of SQLiteGame, do so only in a separate
function call to Initialise().

This allows to construct subclass instances before the storage location
is known; the storage can then be passed later on from a "DefaultMain"
like method.