Refactor GameLogic initialisation.
This refactors and unifies the initialisation of GameLogic instances
based on the context of a Game (mainly after the RPC connection to
Xaya Core has been set up).

Instead of having two separate methods to set the chain and game ID,
both of them are now set by one call to InitialiseGameContext.  In the
future, that can also have more arguments (e.g. the RPC connection itself,
so the game logic has access e.g. to verifymessage).

Also changes the code so that only one initialisation is possible, and
no resetting of the RPC connection of a Game instance is allowed.  Allowing
reconnection is of no real use in practice and just makes things more
complicated than necessary.