Support verifymessage in libxayagame.
This change exposes the RPC client from GameLogic, and adds functionality
to call Xaya Core's verifymessage RPC through that.  With that, games can
verify messages signed by addresses.

This is useful in a couple of situations.  For instance, xid already does
so, and will be changed to use this function now.  Also game-channel games
will need this extensively in various places (including the core GSP).