Initialise libgamechannel with database schema.
This sets up a new library libgamechannel, which will hold the framework
code for building game-channel games.  For now, we define a basic
ChannelGame subclass of SQLiteGame and define a basic database schema
for holding the data of channels in the on-chain state.

Also includes a small change to SQLiteGame which allows exposing the
database handle outside of the callbacks.  This is only intended for
unit testing.