Define interface for channel rules.
This adds a new file, which defines the main interface that games have
to implement to tell the game-channels framework about the rules of the
game on a channel.  ChannelGame instances need to provide access to an
implementation of this interface.

Also moves BoardState from database.hpp there, which is a more proper
place as the interface has less dependencies and deals with states.

Finally, this defines a trivial set of game rules for a simple game that
will be used in unit tests.