Introduce "turn count" concept.
This adds a new function that BoardRules instances have to provide:
Extracting a "turn count" value from the state.  This value will be
used to determine what state is later, e.g. for determining how two
conflicting disputes should be processed.

Also the AdditionRules test game and the tests using it are updated
to implement a turn count.