Implement processing of dispute requests.
The new method ChannelGame::ProcessDispute handles a dispute request
for a given channel.  It verifies that the request is valid, and if
it is, sets the disputed state accordingly.

This method can be used from move processing of concrete games as needed
to implement the on-chain GSP for channel games.