Move gamechannel protos to "proto" namespace.
Define the gamechannel protocol buffers in package "xaya.proto",
so that they are in the xaya::proto C++ namespace.  This separates
them off from the "real" C++ classes, so that it is immediately apparent
what is a protocol buffer.

This is also consistent now with the protocol buffers for mover.