Refactor BoardRules structure.
Basing all methods that inspect a BoardState directly on BoardRules means
that we have to parse encoded data (e.g. protocol buffers) on every method
call.  It also means that all methods need to handle invalid (unparseable)
data in some way.

Instead of that, this change refactors the interface.  Now, BoardRules just
has one method, which parses a state and returns a subclass for another
interface that represents a parsed state.  This class then allows for
extracting data.  This avoids (some) of the unnecessary protocol buffer
deserialisation operations and also puts the check for invalid data into
a single, well-defined place.