Disallow dispute for 'no turn' situation.
This adds another rule for dispute processing, so that it is not possible
to file a dispute against a 'no turn' situation.  This could just cause
confusion, since a dispute essentially means the player whose turn it is
is not responding; that makes no sense for a 'no turn' state.

On the other hand, it is of course possible to *resolve* a dispute to
a 'no turn' state.