Utility templates for board states as protos.
This introduces two utility template classes, ProtoBoardState and
ProtoBoardRules.  They are implementations of ParsedBoardState and
BoardRules, respectively, for situations where the encoded states and
moves are serialised protocol buffers.

The serialisation/deserialisation is handled by the classes, so that
concrete games only have to implement their logic for the actual protocol
buffer instances.