Grid class for holding a position.
The new Grid class holds a bit for every position on our board.  This is
used to encode positions of ships, already targeted coordinates and
the answers of players to shots.

For now, the class just implements a basic bit field.  In the future, the
source file will also have functions for validating that a position is
valid by itself (right number of ships, not touching, ...) and whether
a revealed position of ships matches previous answers of the player.