Include channel ID and topic in data signatures.
When verifying signatures of channel participants for certain data,
include also the channel ID and a "topic" (describing what the data is)
in the hash that gets signed.

By doing that, we make sure that no replay attacks against other channels
can be made, and also not e.g. signed states can be mistaken as signed
winner statements or something like that.