Add TurnCount implementation for ships.
Implement TurnCount in the ships board rules.  The tests that apply
moves are extended to verify that the turn count always increments by
exactly one for each turn.  (This is not strictly necessary to be the
case, but we can do it for ships easily and thus it simplifies things.)