Define database schema for ships on-chain state.
Implement SetupSchema for the Xayaships on-chain GSP, defining the
database schema for the global game state (just tracking game stats
per Xaya name).

To test the schema, this also creates a test fixture with an in-memory
database and associated ShipsLogic.  That will be useful for other tests
in the future as well.