Allow disputes for same-turn non-dispute states.
If there is a non-disputed state on-chain, allow a dispute of the same
turn count (i.e. typically the same state) to be filed nevertheless.
Otherwise a player who just resolved a dispute wouldn't be able to file
one if now the opponent stops responding.

All other situations remain unchanged:  Disputes on disputes as well as
resolutions (on disputes or resolutions) cannot be filed for same turn
states, and disputes cannot be filed with an *earlier* turn count either.

This fixes https://github.com/xaya/libxayagame/issues/51.