Gracefully handle single-participant channels.
Extend the logic in board.cpp so that single-participant channels (i.e.
waiting for a second player to join) are handled fine.  For them every
state is considered as valid, no-turn and with turn count 0.

By doing that, we make sure that those channels can be properly returned
also in the game-state JSON data.