Implement ShipsLogic::GetStateAsJson.
Add a proper implementation of ShipsLogic::GetStateAsJson.  It returns
the game stats from the database and all open channels through the
game-channel framework's gamestatejson functionality.

For each open channel, we also add the current phase as string to
the result JSON; this exposes our GetPhase logic to frontends, so that
they do not have to re-implement it themselves.