Implement rolling state for game channels.
This adds the RollingState class to the game-channels framework.  It has
the ability to keep track of a "current state" and a corresponding state
proof for all known reinitialisation states of a channel, updating those
as new on- or off-chain updates become known.

An instance of this class will be one of the main parts of the channel
management logic, taking care of the actual game states.

This implements https://github.com/xaya/libxayagame/issues/53.