Only allow same-turn disputes for equal states.
Instead of allowing disputes to be filed with the same turn count
as previous resolutions, only allow them if the state is actually
the same.  This also resolves the issue from
https://github.com/xaya/libxayagame/issues/51, but does so without the
weird "dispute counts as later than resolution" rule.

With this change, we are guaranteed that the on-chain states for
some channel are strictly increasing in turn count.  Filing same-height
disputes does not change the state at all, it just marks the channel
as disputed in contrast to the previous state.