Keep full state proofs in the on-chain state.
As described in https://github.com/xaya/libxayagame/issues/54, we should
keep both a "frozen" reinitialisation state and *full state proofs*
in the on-chain game state (rather than just a latest on-chain state).

Having state proofs makes sure that they remain valid independent of
the on-chain context, so that reorgs or changes to the on-chain state
will not invalidate the state proof of some player (which could open
avenues for cheating).

This change updates the database storage.  After that, we also need to
modify the state proof logic itself to rely on the reinit state only,
and update the game-state JSON for channels to return the full state
proof and reinit state in addition to the latest state.