Introduce uint256::GetBinaryString.
There are some places in the code that need to convert a uint256 to
a binary std::string.  Based on GetBlob(), this is a bit ugly as it
needs a reinterpret_cast.  This change adds a simple utility method
uint256::GetBinaryString that does exactly that, and uses it in the
various places.