Add reinitialisation ID to channel metadata.
This introduces the concept of a "reinitialisation ID", which is added
to the metadata of a channel.  Whenever the channel's metadata changes
through some on-chain action (e.g. someone joins the channel) and thus
the channel game is reinitialised, this ID should be updated.

In a follow-up, we will make sure this ID gets hashed into signatures
(like the channel ID itself) to prevent potential replay attacks.  The ID
can also be used to verify for what "revision" of a channel some state
proof is and thus apply it correctly when managing off-chain data relayed
for a channel.