Extended game-state JSON for channels.
This extends the game-state JSON returned for channels.  In particular,
the new "reinit id" is returned in the metadata, and also the full metadata
protocol buffer as base64-encoded serialised bytes.

Furthermore, we now return both the current state (as previously) and
also the reinit state.  Also the state proof corresponding to the
current state is returned as base64-encoded protocol buffer.