Use reinit state as basis for state proofs.
Instead of using the latest on-chain state as "basis" for verifying
state proofs, use the channel's reinit state.  These states are fixed,
so that state proofs cannot be invalidated easily (for more details,
see also https://github.com/xaya/libxayagame/issues/54).

Note that this also means that "most" state proofs will need to be
fully signed, and only proofs for the very first states in a game can
ever have fewer signatures because of being based on the initial state.