Always return encoded board state in JSON.
This changes the JSON format in which board states are returned slightly.
The main change is that we now always return the base64-encoded board
state, independently of the ParsedBoardState::ToJson method (which did
it previously for protoboards).

This makes sense, since the full encoded board state is useful in any
case, independently of protocol buffers.  This will allow usage of
that data e.g. for general channel-management code that is not related
to any of the specific games or even protocol-buffer-based games.