Interface for MoveSender and broadcast.
Define the interfaces that are used by the upcoming ChannelManager to
send on- and off-chain moves.  They will be implemented at a later stage
using the real-time communication system and Xaya Core name_update's, but
for now they are just useful for implementing the ChannelManager itself
and unit tests (where the interfaces are mocked).