Utility method to extract endstate from proof.
The new function UnverifiedProofEndState can efficiently extract the
end state from a StateProof without verifying it.  Its logic was inlined
already in some places, but it is generally useful in some situations
(e.g. when a proof comes from the on-chain GSP or has been verified
before already), so it makes sense to have it well-defined and tested.