Add ChannelManager with resolution handling.
This creates the new ChannelManager class.  It wraps a RollingState
instance to keep track of the best known board states, but it also tracks
the state of disputes and is able to automatically attempt resolutions
if a dispute is filed against the player but we already have a better
current state.

In the future, this class will also be able to submit local moves
(and coordiante signing the resulting state and building a state proof
out of it) as well as provide access to its state by functions that
will be the basis of the RPC interface.