Add RPC stub for Xaya Core wallet.
Define a subset of the RPC interface of the Xaya Core wallet for use with
libjson-rpc-cpp.  In particular, the interface supports getaddressinfo (to
check if an address is owned), signmessage and name_update.  With this, we
can cover the existing usage in xid and the minimal usage requirements
for the upcoming channel-management code.