Code for building / extending state proofs.
This defines a new function ExtendStateProof, which is able to add a move
to a given state proof.  It takes care of signing the new state by the
correct player (assuming it is given a wallet that has signing keys), as
well as "minimising" the resulting state proof to just the required number
of transitions after the new one.