Proper shutdown sequence for ChannelManager.
Add a function to the ChannelManager for stopping updates.  This is used
when shutting it down so that we can properly ensure that all threads
blocked on WaitForChange are woken up and no more threads will start
blocking in this function.

This has a similar function to shutting down the ZMQ subscriber for a
Game instance.