Versioned ChannelManager::WaitForChange.
In the spirit of https://github.com/xaya/libxayagame/pull/63, add
versioning to the state of a ChannelManager.  Whenever we update it,
the version is incremented.  It is returned from ToJson(), and
WaitForChange supports passing in a known version so that it can
return without blocking if the channel already is in a newer version.

This avoids potential race conditions if the state changes between
two calls to WaitForChange.