Add getchannel RPC method for channel GSPs.
This defines a "standard RPC server" for channel GSPs, which extends
GameRpcServer by a getchannel method that can extract just the JSON
data for one of the channels.

This method will be useful for feeding on-chain updates into a channel
daemon, i.e. https://github.com/xaya/libxayagame/issues/58.

We also apply the new RPC server for shipsd, and add an RPC test
for it to the ships integration tests.