Return "" instead of null from waitforchange.
When waitforchange returns and the GSP does not yet have any state,
return "" instead of null.  This allows us to change the return type
for waitforchange in the declaration to a string, and ensure that we
always return a string.

It seems that libjson-rpc-cpp has issues when the result of a method
can be of different types; always returning a string mitigates this.