Add timeout for WaitForChange.
Instead of blocking for an unlimited amount of time, make sure that
the two WaitForChange methods (in Game and ChannelManager) return after
at most five seconds.

They are already permitted to return spuriously, so clients need to be
able to handle that anyway.  But returning every five seconds makes sure
that we do not collect a large number of blocked threads.

This implements https://github.com/xaya/libxayagame/issues/66.