Actual implementation for MoveSender.
This replaces the MoveSender interface with a class that actually
implements sending moves through name_update's with a Xaya wallet
(https://github.com/xaya/libxayagame/issues/60).

We add a new interface OpenChannel, which will hold the game-specific
data and logic for a channel daemon.  For now, it has functions to
construct the (game-specific) moves for opening disputes and sending
resolutions on-chain.