General logic for broadcast system.
This replaces the OffChainBroadcast interface with a class that will
be the basis for real broadcast implementations.  It has general logic
that is shared, and defines the interface that concrete implementations
of a broadcast system have to implement.

This is the main step for https://github.com/xaya/libxayagame/issues/59
(with the remaining piece being implementing a concrete broadcast
system based on a server).