Share code for test fixture with ChannelManager.
Extract code to set up a test fixture with ChannelManager and predefined
addresses (as well as mock expectations for validating and creating
signatures with them).  That code was more or less duplicated between
channelmanager_tests and chaintochannel_tests before, and is now shared
properly between those two test suites.

It will also be useful for testing the upcoming implementation of
the general broadcasting logic.