Implement RPC-server-based broadcast system.
This adds a concrete implementation of an off-chain broadcast system,
which is based on a JSON-RPC server that routes and distributes the
messages sent to it by function calls.

The system is not yet fully ready for production, but it may be so
with a few tweaks.  It is definitely ready for use in integration tests
as well as "real" tests in somewhat controlled environments.